==> pythonProject5/notebooks/1.0-initial-data-analysis.ipynb <==
## INITIAL DATA ANALYSIS 

import pandas as pd

#read data
df = pd.read_csv('../data/raw_filtered.tsv', sep='\t')

df.head()

df.info()

## [DATA PREPROCESSING]()

#### [TEXT CLEANING AND NORMALIZATION]()

import re
import nltk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize
from nltk.corpus import wordnet
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer

nltk.download('stopwords')
nltk.download('punkt')
nltk.download('wordnet')

stop_words = set(stopwords.words('english'))
lemmatizer = WordNetLemmatizer()

def clean_text(text):
    text = text.lower()
    text = re.sub(r'[^a-zA-Z0-9\s]', ' ', text)
    text = re.sub(r'\s+', ' ', text)
    text = text.strip()
    return text

def remove_stopwords(text):
    tokens = word_tokenize(text)
    filtered_tokens = [token for token in tokens if token not in stop_words]
    filtered_text = ' '.join(filtered_tokens)
    return filtered_text

def lemmatize_text(text):
    tokens = word_tokenize(text)
    lemmatized_tokens = [lemmatizer.lemmatize(token, pos='v') for token in tokens]
    lemmatized_text = ' '.join(lemmatized_tokens)
    return lemmatized_text

def preprocess_text(text):
    text = clean_text(text)
    text = remove_stopwords(text)
    text = lemmatize_text(text)
    return text

df['reference'] = df['reference'].apply(preprocess_text)
df['translation'] = df['translation'].apply(preprocess_text)

# save as "cleaned" file
df.to_csv('../data/inter_cleaned.tsv', sep='\t', index=False)

cleaned = pd.read_csv('../data/inter_cleaned.tsv', sep='\t')

cleaned.drop("Unnamed: 0", axis=1, inplace=True)

cleaned.head()

#### [Handling NONE]()

# delete rows with NONE
cleaned = cleaned[cleaned['reference'].notna()]
cleaned = cleaned[cleaned['translation'].notna()]

# show length of reference and translation
cleaned['ref_len'] = cleaned['reference'].apply(lambda x: len(x.split()))
cleaned['trn_len'] = cleaned['translation'].apply(lambda x: len(x.split()))


cleaned.head()

#### [HANDLING NUMERICAL VALUES]()

# replace numerical values with token "num"
cleaned['reference'] = cleaned['reference'].str.replace('\d+', 'num')
cleaned['translation'] = cleaned['translation'].str.replace('\d+', 'num')

#cleaned = cleaned.drop(['ref_len', 'trn_len'], axis=1)

cleaned.head()

# if translation toxicity is higher than reference toxicity, then swap them
for i in range(len(cleaned)):
    if cleaned.iloc[i]['ref_tox'] < cleaned.iloc[i]['trn_tox']:
        cleaned.iloc[i]['reference'], cleaned.iloc[i]['translation'] = cleaned.iloc[i]['translation'], cleaned.iloc[i]['reference']
        cleaned.iloc[i]['ref_tox'], cleaned.iloc[i]['trn_tox'] = cleaned.iloc[i]['trn_tox'], cleaned.iloc[i]['ref_tox']
        
cleaned.head()

# save cleaned data as "cleaned_new.tsv"
cleaned.to_csv('../data/final_cleaned.tsv', sep='\t', index=False)
