==> pythonProject5/notebooks/2.0_classification-model-creation-2.ipynb <==
import pandas as pd
import numpy as np
import torch

from torch.utils.data import DataLoader
from transformers import BertTokenizer, BertForSequenceClassification, BertConfig
from torch.utils.data import DataLoader, TensorDataset
from transformers import T5Tokenizer, T5ForConditionalGeneration, T5Config
from transformers import BertTokenizer, BertModel
from transformers import GPT2LMHeadModel, GPT2Tokenizer
from sklearn.model_selection import train_test_split
from transformers import BertTokenizer, BertForSequenceClassification, BertConfig
from torch.utils.data import DataLoader, TensorDataset

from sklearn.metrics import accuracy_score
from sklearn.metrics import f1_score
from sklearn.metrics import precision_score
from sklearn.metrics import recall_score
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.metrics import confusion_matrix
import torch.nn as nn
import torch.nn.functional as F
from torch.utils.data import TensorDataset, DataLoader
from sklearn.metrics import classification_report

from transformers import BertTokenizer
from transformers import BertForSequenceClassification
from transformers import AdamW
from transformers import get_linear_schedule_with_warmup
from transformers import BertForSequenceClassification
import torch.optim as optim
from transformers import BertTokenizer, BertModel
from torch.nn.utils.rnn import pad_sequence

# read tsv file cleaned_new
df = pd.read_csv('../data/final_cleaned.tsv', sep='\t')
df = df[:1000]
df.head()


### Embeddings:

Convert your tokenized text into numerical representations using word embeddings. You can use pre-trained embeddings (Word2Vec, GloVe, BERT) to capture the semantic meaning of words.

Padding and Sequence Length:
Ensure that all input sequences (toxic text) and output sequences (non-toxic text) have the same length. You may need to pad shorter sequences with a special token (e.g., <PAD>) or truncate longer sequences to a fixed length.

model_name = 'bert-base-uncased'
tokenizer = BertTokenizer.from_pretrained(model_name)
config = BertConfig.from_pretrained(model_name)
model = BertForSequenceClassification.from_pretrained(model_name, num_labels=2)

texts = df['reference'].tolist() + df['translation'].tolist()
toxicity = df['ref_tox'].tolist() + df['trn_tox'].tolist()
toxicity = [1 if x > 0.35 else 0 for x in toxicity]
input_ids = []
attention_masks = []

for text in texts:
    encoded_dict = tokenizer.encode_plus(
        text,
        add_special_tokens=True,
        max_length=64,
        pad_to_max_length=True,
        return_attention_mask=True,
        return_tensors='pt',
    )

    input_ids.append(encoded_dict['input_ids'])
    attention_masks.append(encoded_dict['attention_mask'])

input_ids = torch.cat(input_ids, dim=0)
attention_masks = torch.cat(attention_masks, dim=0)
toxicity = torch.tensor(toxicity)

dataset = TensorDataset(input_ids, attention_masks, toxicity)
train_size = int(0.8 * len(dataset))
val_size = len(dataset) - train_size

train_dataset, val_dataset = torch.utils.data.random_split(dataset, [train_size, val_size])
train_dl = DataLoader(train_dataset, batch_size=32, shuffle=True)
val_dl = DataLoader(val_dataset, batch_size=32, shuffle=True)

device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
model = model.to(device)
optimizer = torch.optim.AdamW(model.parameters(), lr=1e-5)

epochs = 100

for epoch in range(epochs):
    print('Epoch:', epoch + 1)
    model.train()
    total_loss = 0
    for batch in train_dl:
        batch = tuple(t.to(device) for t in batch)
        b_input_ids, b_input_mask, b_labels = batch
        optimizer.zero_grad()
        outputs = model(b_input_ids, token_type_ids=None, attention_mask=b_input_mask, labels=b_labels)
        loss = outputs[0]
        total_loss += loss.item()
        loss.backward()
        optimizer.step()
    avg_train_loss = total_loss / len(train_dl)
    print("Average train loss: {}".format(avg_train_loss))

model.eval()
correct = 0
total = 0

with torch.no_grad():
    for batch in val_dl:
        batch = tuple(t.to(device) for t in batch)
        b_input_ids, b_input_mask, b_labels = batch

        outputs = model(b_input_ids, token_type_ids=None, attention_mask=b_input_mask)
        logits = outputs[0]

        logits = logits.detach().cpu().numpy()
        label_ids = b_labels.to('cpu').numpy()

        pred_flat = np.argmax(logits, axis=1).flatten()
        labels_dlat = label_ids.flatten()

        total += labels_dlat.size
        correct += np.sum(pred_flat == labels_dlat)

print('Accuracy: %d %%' % (100 * correct / total))

def predict_toxicity(text, model):
    encoded_dict = tokenizer.encode_plus(
        text,
        add_special_tokens=True,
        max_length=64,
        pad_to_max_length=True,
        return_attention_mask=True,
        return_tensors='pt',
    )
    input_ids = encoded_dict['input_ids']
    attention_mask = encoded_dict['attention_mask']
    input_ids = input_ids.to(device)
    attention_mask = attention_mask.to(device)
    outputs = model(input_ids, token_type_ids=None, attention_mask=attention_mask)
    logits = outputs[0]
    logits = logits.detach().cpu().numpy()
    pred_flat = np.argmax(logits, axis=1).flatten()
    return pred_flat[0]

print(predict_toxicity('I love you'))
print(predict_toxicity('I hate you'))

print(predict_toxicity('I hate you so much asshole'))

print(predict_toxicity('I love you so much'))

# save model
torch.save(model.state_dict(), '../models/classification_model_NEW.pt')

# load saved model
model_my = BertForSequenceClassification.from_pretrained(model_name, num_labels=2)

model_my.load_state_dict(torch.load('../models/classification_model_NEW.pt'))

def predict_toxicity(text, model):
    encoded_dict = tokenizer.encode_plus(
        text,
        add_special_tokens=True,
        max_length=64,
        pad_to_max_length=True,
        return_attention_mask=True,
        return_tensors='pt',
    )
    input_ids = encoded_dict['input_ids']
    attention_mask = encoded_dict['attention_mask']
    input_ids = input_ids.to(device)
    attention_mask = attention_mask.to(device)
    outputs = model(input_ids, token_type_ids=None, attention_mask=attention_mask)
    logits = outputs[0]
    logits = logits.detach().cpu().numpy()
    pred_flat = np.argmax(logits, axis=1).flatten()
    return pred_flat[0]

model_my.to(device)
predict_toxicity('I love you', model_my)

import gc

gc.collect()
